Added transition between encounter and choose encounter state. fixed monster hp bar (need to implement same fix for player HP) note:changed some player data values to float and some labels must be changed as a result.

+++ Project/PlayerData.h
@@ -14,8 +14,8 @@
 @property float MaxHP;
 @property float HP;
 @property int AttackPower;
-@property int Coins;
-@property int Score;
+@property float Coins;
+@property float Score;
 
 + (PlayerData *) sharedInstance;
